# Optimize instruction fetch and decoding