--- a/Makefile
+++ b/Makefile
@@ -150,7 +150,7 @@ export CXX=g++
 
 endif
 
-GENERATED_FILES= uarch/uarch-pristine-hash.c uarch/uarch-pristine-ram.c src/machine-c-version.h
+GENERATED_FILES= uarch/uarch-pristine-hash.c uarch/uarch-pristine-ram.c src/machine-c-version.h src/interpret-jump-table.h
 ADD_GENERATED_FILES_DIFF= add-generated-files.diff
 
 all: source-default
@@ -240,11 +240,14 @@ lint-% check-format-% format-% check-format-lua-% check-lua-% format-lua-%:
 source-default:
 	@eval $$($(MAKE) -s --no-print-directory env); $(MAKE) -C $(SRCDIR)
 
-uarch: $(SRCDIR)/machine-c-version.h
+uarch: $(SRCDIR)/machine-c-version.h $(SRCDIR)/interpret-jump-table.h
 	@eval $$($(MAKE) -s --no-print-directory env); $(MAKE) -C uarch
 
 $(SRCDIR)/machine-c-version.h:
 	@eval $$($(MAKE) -s --no-print-directory env); $(MAKE) -C $(SRCDIR) machine-c-version.h
+
+$(SRCDIR)/interpret-jump-table.h:
+	@eval $$($(MAKE) -s --no-print-directory env); $(MAKE) -C $(SRCDIR) interpret-jump-table.h
 
 build-emulator-builder-image:
 	docker build $(DOCKER_PLATFORM) --build-arg DEBUG=$(debug) --build-arg COVERAGE=$(coverage) --build-arg SANITIZE=$(sanitize) --target builder -t cartesi/machine-emulator:builder -f Dockerfile .
@@ -278,6 +281,7 @@ copy:
 	docker create --name uarch-ram-bin $(DOCKER_PLATFORM) $(DEBIAN_IMG)
 	docker cp uarch-ram-bin:/usr/src/emulator/$(DEB_FILENAME) .
 	docker cp uarch-ram-bin:/usr/src/emulator/src/machine-c-version.h .
+	docker cp uarch-ram-bin:/usr/src/emulator/src/interpret-jump-table.h .
 	docker cp uarch-ram-bin:/usr/src/emulator/uarch/uarch-ram.bin .
 	docker cp uarch-ram-bin:/usr/src/emulator/uarch/uarch-pristine-ram.c .
 	docker cp uarch-ram-bin:/usr/src/emulator/uarch/uarch-pristine-hash.c .
@@ -390,4 +394,3 @@ $(ADD_GENERATED_FILES_DIFF): $(GENERATED_FILES)
 
 .PHONY: help all submodules doc clean distclean src luacartesi hash uarch \
 	create-generated-files-patch $(SUBDIRS) $(SUBCLEAN)
-