--- a/src/.gitignore
+++ b/src/.gitignore
@@ -8,3 +8,4 @@ compile_flags.txt
 coverage*
 jsonrpc-discover.cpp
 machine-c-version.h
+interpret-jump-table.h