--- a/src/machine-state.h
+++ b/src/machine-state.h
@@ -51,13 +51,14 @@ struct machine_state {
 
     // The following state fields are very hot,
     // and are carefully ordered to have better data locality in the interpreter loop.
+    // The X registers are the very first to optimize access of registers in the interpreter.
+    std::array<uint64_t, X_REG_COUNT> x{}; ///< Register file
     uint64_t mcycle{};                     ///< CSR mcycle.
     uint64_t pc{};                         ///< Program counter.
-    std::array<uint64_t, X_REG_COUNT> x{}; ///< Register file.
     uint64_t fcsr{};                       ///< CSR fcsr.
     std::array<uint64_t, F_REG_COUNT> f{}; ///< Floating-point register file.
 
-    uint64_t icycleinstret{}; ///< CSR icycleinstret.
+    uint64_t iprv{}; ///< Privilege level (Cartesi-specific).
 
     uint64_t mstatus{};  ///< CSR mstatus.
     uint64_t mtvec{};    ///< CSR mtvec.
@@ -84,8 +85,8 @@ struct machine_state {
     uint64_t senvcfg{};    ///< CSR senvcfg.
 
     // Cartesi-specific state
-    uint64_t ilrsc{}; ///< For LR/SC instructions (Cartesi-specific).
-    uint64_t iprv{};  ///< Privilege level (Cartesi-specific).
+    uint64_t ilrsc{};         ///< For LR/SC instructions (Cartesi-specific).
+    uint64_t icycleinstret{}; ///< CSR icycleinstret (Cartesi-specific).
     struct {
         uint64_t X{}; ///< CPU has yielded with automatic reset (Cartesi-specific).
         uint64_t Y{}; ///< CPU has yielded with manual reset (Cartesi-specific).