This optimizes our RISC-V instruction decoder by using big jump tables, through [token threading](https://en.wikipedia.org/wiki/Threaded_code#Token_threading), to the point that decoding takes just 2 instructions for most RISC-V instructions, even compressed ones. And overall the FENCE instruction takes 12 host instructions in GCC  AMD64, and Clang ARM64.

Here is the GCC x86_64 trace as proof:

```asm
//// FENCE GCC x86_64 (2/12 instructions)
// increment mcycle (3 instructions)
=> 0x7ffff7a2e98c <loop+28108>:   add    $0x1,%r15                     // mcycle += 1
=> 0x7ffff7a2e990 <loop+28112>:   cmp    %r13,%r15                     // mcycle < mcycle_tick_end
=> 0x7ffff7a2e993 <loop+28115>:   jae    0x7ffff7a2f230 <loop+30320>   // -> break loop
// fetch (5 instructions)
=> 0x7ffff7a2e999 <loop+28121>:   mov    %r10,%rbx                     // pc
=> 0x7ffff7a2e99c <loop+28124>:   xor    %rbp,%rbx                     // pc ^ fetch_vaddr_page
=> 0x7ffff7a2e99f <loop+28127>:   cmp    $0xffd,%rbx                   // check fetch page
=> 0x7ffff7a2e9a6 <loop+28134>:   ja     0x7ffff7a27d00 <loop+320>     // -> miss fetch
=> 0x7ffff7a2e9ac <loop+28140>:   mov    (%r14,%rbp,1),%ebx            // insn = *(uint32_t*)(pc + fetch_vh_offset)
// decode (2 instructions)
=> 0x7ffff7a2e9b0 <loop+28144>:   movzwl %bx,%ecx                      // insn & 0b1111111111111111
=> 0x7ffff7a2e9b3 <loop+28147>:   jmp    *(%r11,%rcx,8)                // -> jump to instruction
// execute (2 instructions)
=> 0x7ffff7a2ea3b <loop+28283>:   add    $0x4,%rbp                     // pc += 4
=> 0x7ffff7a2ea3f <loop+28287>:   jmp    0x7ffff7a2e98c <loop+28108>   // -> jump to loop begin
```

And the Clang arm64:

```asm
//// FENCE Clang arm64 (2/12 instructions)
// increment mcycle
=> 0xfffff7b8a328 <loop+4568>:    add x25, x25, $0x1
=> 0xfffff7b8a32c <loop+4572>:    cmp x25, x27
=> 0xfffff7b8a330 <loop+4576>:    b.cs    0xfffff7b8e7a8 <loop+22104>
// fetch
=> 0xfffff7b8a334 <loop+4580>:    eor x19, x20, x28
=> 0xfffff7b8a338 <loop+4584>:    cmp x19, $0xffd
=> 0xfffff7b8a33c <loop+4588>:    b.hi    0xfffff7b89264 <loop+276>
=> 0xfffff7b8a340 <loop+4592>:    ldr w19, [x20, x22]
// decode
=> 0xfffff7b8a344 <loop+4596>:    and w10, w19, $0xffff
=> 0xfffff7b8a348 <loop+4600>:    ldr x16, [x24, x10, lsl $3]
=> 0xfffff7b8a34c <loop+4604>:    br  x16
// execute
=> 0xfffff7b8dde8 <loop+19608>:   add x20, x20, $0x4
=> 0xfffff7b8ddec <loop+19612>:   b   0xfffff7b8a328 <loop+4568>
```

In emulator v0.18.x the trace for same FENCE RISC-V instruction took about 40 x86_64 instructions.

Overall the performance varies between 1.2x up to 2x speedup across many benchmarks relative to emulator v0.18.1, here is results for many benchmarks with `stress-ng` :

## Benchmarks

| Times faster | Benchmark |
|-------------|---------------------------------------------------------|
| 2.56 ± 0.03 | stress-ng --no-rand-seed --syscall 1 --syscall-ops 4000 |
| 2.15 ± 0.02 | stress-ng --no-rand-seed --cpu 1 --cpu-method loop --cpu-ops 400 |
| 1.95 ± 0.00 | stress-ng --no-rand-seed --cpu 1 --cpu-method fibonacci --cpu-ops 400 |
| 1.94 ± 0.01 | stress-ng --no-rand-seed --cpu 1 --cpu-method int64 --cpu-ops 400 |
| 1.90 ± 0.01 | stress-ng --no-rand-seed --memcpy 1 --memcpy-ops 50 |
| 1.88 ± 0.02 | stress-ng --no-rand-seed --crypt 1 --crypt-method SHA-256 --crypt-ops 400000 |
| 1.87 ± 0.01 | stress-ng --no-rand-seed --qsort 1 --qsort-ops 5 |
| 1.83 ± 0.01 | stress-ng --no-rand-seed --memrate 1 --memrate-bytes 2M --memrate-ops 200 |
| 1.82 ± 0.03 | stress-ng --no-rand-seed --hash 1 --hash-ops 40000 |
| 1.75 ± 0.00 | stress-ng --no-rand-seed --heapsort 1 --heapsort-ops 3 |
| 1.72 ± 0.01 | stress-ng --no-rand-seed --zlib 1 --zlib-ops 20 |
| 1.66 ± 0.00 | stress-ng --no-rand-seed --matrix 1 --matrix-method mult --matrix-ops 20000 |
| 1.49 ± 0.02 | stress-ng --no-rand-seed --hdd 1 --hdd-ops 2000 |
| 1.41 ± 0.00 | stress-ng --no-rand-seed --fp 1 --fp-method floatadd --fp-ops 1000 |
| 1.33 ± 0.01 | stress-ng --no-rand-seed --fma 1 --fma-ops 40000 |
| 1.24 ± 0.01 | stress-ng --no-rand-seed --trig 1 --trig-ops 50 |
| 1.16 ± 0.01 | stress-ng --no-rand-seed --fork 1 --fork-ops 1000 |
| 1.14 ± 0.01 | stress-ng --no-rand-seed --malloc 1 --malloc-ops 40000 |

You can see 1.94 a speedup for integer operations. Notably I am able to reach GHz speed for some simple integer arithmetic benchmarks, with the interpreter being only 10~20x slower than host native.

The table of benchmarks were created by running `hyperfine` and `stress-ng`, for example:

```sh
$ hyperfine -w 1 'cartesi-machine -- stress-ng --no-rand-seed --cpu 1 --cpu-method loop --cpu-ops 400' '/usr/bin/cartesi-machine -- stress-ng --no-rand-seed --cpu 1 --cpu-method loop --cpu-ops 400'
Benchmark 1: cartesi-machine -- stress-ng --no-rand-seed --cpu 1 --cpu-method loop --cpu-ops 400
  Time (mean ± σ):      2.225 s ±  0.021 s    [User: 2.213 s, System: 0.010 s]
  Range (min … max):    2.197 s …  2.257 s    10 runs
 
Benchmark 2: /usr/bin/cartesi-machine -- stress-ng --no-rand-seed --cpu 1 --cpu-method loop --cpu-ops 400
  Time (mean ± σ):      4.615 s ±  0.041 s    [User: 4.602 s, System: 0.009 s]
  Range (min … max):    4.561 s …  4.682 s    10 runs
 
Summary
  cartesi-machine -- stress-ng --no-rand-seed --cpu 1 --cpu-method loop --cpu-ops 400 ran
    2.07 ± 0.03 times faster than /usr/bin/cartesi-machine -- stress-ng --no-rand-seed --cpu 1 --cpu-method loop --cpu-ops 400
```

## PREVIOUS PR ITERATION COMMENTS

This a micro optimization at x86_64 assembly level of the instruction fetch+decode hot path. In summary this PR should save about 22 x86_64 instructions from every interpreter hot loop iteration. This optimization does not apply only to x86_64, but all architectures should benefit from.

## Baseline

First I generated a hot trace of subsequent `FENCE.I` instruction calls. I choose this instruction because it is the most simple instruction, it basically does nothing, it's the ideal instruction to measure instruction fetch ovearhead. This was the trace for one iteration:

```asm
// mcycle check
| 0x7ffff7b88330 <interpret_loop+432>    add    $0x1,%r14                │ ++mcycle
│ 0x7ffff7b88334 <interpret_loop+436>    cmp    %r11,%r14                │ mcycle < mcycle_tick_end
│ 0x7ffff7b88337 <interpret_loop+439>    jae    0x7ffff7b88420           │ -> break interpret hot loop
// fetch
│ 0x7ffff7b8833d <interpret_loop+445>    mov    %r15,%rbx                │ pc
│ 0x7ffff7b88340 <interpret_loop+448>    and    $0xfffffffffffff000,%rbx │ vaddr_page = pc & ~PAGE_OFFSET_MASK
│ 0x7ffff7b88347 <interpret_loop+455>    cmp    %r12,%rbx                │ vaddr_page == fetch_vaddr_page
│ 0x7ffff7b8834a <interpret_loop+458>    jne    0x7ffff7b88728           │ -> miss fetch cache
│ 0x7ffff7b88350 <interpret_loop+464>    lea    0x0(%r13,%r15,1),%rax    │ hptr = pc + fetch_vh_offset
│ 0x7ffff7b88355 <interpret_loop+469>    mov    %r15,%rdx                │ pc
│ 0x7ffff7b88358 <interpret_loop+472>    not    %rdx                     │ ~pc
│ 0x7ffff7b8835b <interpret_loop+475>    test   $0xffe,%edx              │ ((~pc & PAGE_OFFSET_MASK) >> 1) == 0
│ 0x7ffff7b88361 <interpret_loop+481>    je     0x7ffff7b88760           │ -> cross page boundary
│ 0x7ffff7b88367 <interpret_loop+487>    mov    (%rax),%r9d              │ insn = *(uint32_t*)(hptr)
│ 0x7ffff7b8836a <interpret_loop+490>    mov    %rbx,%r12                │ fetch_vaddr_page = vaddr_page
// decoding: check if is a compressed instruction
│ 0x7ffff7b8836d <interpret_loop+493>    mov    %r9d,%eax                │ insn
│ 0x7ffff7b88370 <interpret_loop+496>    not    %eax                     │ ~insn
│ 0x7ffff7b88372 <interpret_loop+498>    test   $0x3,%al                 │ (~insn & 3) > 0
│ 0x7ffff7b88374 <interpret_loop+500>    jne    0x7ffff7b882b0           │ -> decode compressed instruction
// decoding: decode fence.i uncompressed instruction
│ 0x7ffff7b8837a <interpret_loop+506>    mov    %r9d,%eax                │
│ 0x7ffff7b8837d <interpret_loop+509>    and    $0x707f,%eax             │
│ 0x7ffff7b88382 <interpret_loop+514>    cmp    $0x3023,%eax             │
│ 0x7ffff7b88387 <interpret_loop+519>    je     0x7ffff7b8a868           │
│ 0x7ffff7b8838d <interpret_loop+525>    ja     0x7ffff7b88500           │
│ 0x7ffff7b88393 <interpret_loop+531>    cmp    $0x101b,%eax             |
│ 0x7ffff7b88398 <interpret_loop+536>    je     0x7ffff7b8a820           │
│ 0x7ffff7b8839e <interpret_loop+542>    ja     0x7ffff7b887d0           │
│ 0x7ffff7b883a4 <interpret_loop+548>    cmp    $0x3b,%eax               │
│ 0x7ffff7b883a7 <interpret_loop+551>    je     0x7ffff7b8a638           │
│ 0x7ffff7b883ad <interpret_loop+557>    ja     0x7ffff7b88b68           │
│ 0x7ffff7b88b68 <interpret_loop+2536>   cmp    $0x1003,%eax             │
│ 0x7ffff7b88b6d <interpret_loop+2541>   je     0x7ffff7b8a494           │
│ 0x7ffff7b88b73 <interpret_loop+2547>   ja     0x7ffff7b89450           │
| 0x7ffff7b89450 <interpret_loop+4816>   cmp    $0x1013,%eax             │
│ 0x7ffff7b89455 <interpret_loop+4821>   je     0x7ffff7b8a6d0           │
│ 0x7ffff7b8945b <interpret_loop+4827>   cmp    $0x1017,%eax             │
│ 0x7ffff7b89460 <interpret_loop+4832>   je     0x7ffff7b8a71c           │
│ 0x7ffff7b89466 <interpret_loop+4838>   cmp    $0x100f,%eax             │
│ 0x7ffff7b8946b <interpret_loop+4843>   jne    0x7ffff7b8ab45           │
// execute
│ 0x7ffff7b89471 <interpret_loop+4849>   add    $0x4,%r15                │ pc += 4
│ 0x7ffff7b89475 <interpret_loop+4853>   jmp    0x7ffff7b88330           │ -> jump to begin
```

This trace keeps looping in x86_64. We can see that in optimal conditions it takes exactly 40 x86_64 instructions to execute one `FENCE.I` in this trace, where:
* mcycle check: 3 instructions
* fetch: 11 instructions
* decoding: 24 instructions
* execution: 2 instructions

I usually say that the cartesi machine is about 30~40 times slower than native, if we think about the ratio 40:1 in this trace, this is very close to what I usually say. If  we can get this trace to execute with fewer x86_64 instruction, we can also get the cartesi machine interpreter to be faster for all instructions (not only this one).

If we look closely in the fetch, there are these two branches:
```asm
│ 0x7ffff7b8834a <interpret_loop+458>    jne    0x7ffff7b88728           │ -> miss fetch cache
│ 0x7ffff7b88361 <interpret_loop+481>    je     0x7ffff7b88760           │ -> cross page boundary
```

My idea was to come up with a single branch that could test both conditions in the fetch loop, simplifying to just one branch, so I could save some instructions. 

This is the benchmark for baseline.
```
$ lua bench-insns.lua
RISC-V Privileged Memory-management             4.205 MIPS   3250.0 ucycles
RISC-V Privileged Interrupt-management        495.199 MIPS    257.0 ucycles
RV64I - Base integer instruction set          574.503 MIPS    263.8 ucycles
RV64M - Integer multiplication and division   575.371 MIPS    312.2 ucycles
RV64A - Atomic instructions                   431.893 MIPS    304.2 ucycles
RV64F - Single-precision floating-point       218.556 MIPS    489.3 ucycles
RV64D - Double-precision floating-point       214.964 MIPS   1773.7 ucycles
RV64Zicsr - Control and status registers      289.061 MIPS    328.3 ucycles
RV64Zicntr - Base counters and timers         343.063 MIPS    284.3 ucycles
RV64Zifence - Instruction fetch fence         792.319 MIPS    246.0 ucycles

$ hyperfine -w 10 -m 32 "cartesi-machine dhrystone 2000000"
Benchmark 1: cartesi-machine dhrystone 2000000
  Time (mean ± σ):      1.333 s ±  0.007 s    [User: 1.326 s, System: 0.006 s]
  Range (min … max):    1.321 s …  1.356 s    32 runs
```


## Round 1 - Optimize fetch

After some thinking I come up with the changes presented in the PR to optimize instruction fetch, which generates the following new trace:

```asm
// mcycle check
|   0x7ffff7b873e0 <interpret_loop+432>    add    $0x1,%r15               │ ++mcycle
│   0x7ffff7b873e4 <interpret_loop+436>    cmp    %r10,%r15               │ mcycle < mcycle_tick_end
│   0x7ffff7b873e7 <interpret_loop+439>    jae    0x7ffff7b874b0          │ -> break interpret hot loop
// fetch
│   0x7ffff7b873ed <interpret_loop+445>    mov    %rbp,%rax               │ pc
│   0x7ffff7b873f0 <interpret_loop+448>    xor    %r14,%rax               │ pc ^ fetch_vaddr_page
│   0x7ffff7b873f3 <interpret_loop+451>    cmp    $0xffd,%rax             │ (pc ^ fetch_vaddr_page) < (PMA_PAGE_SIZE - 2)
│   0x7ffff7b873f9 <interpret_loop+457>    ja     0x7ffff7b877d8          │ -> miss fetch cache
│   0x7ffff7b873ff <interpret_loop+463>    mov    0x0(%rbp,%r13,1),%ebx   │ insn = *(uint32_t*)(pc + fetch_vh_offset)
// decoding: check if is a compressed instruction
│   0x7ffff7b87404 <interpret_loop+468>    mov    %ebx,%eax               │ insn
│   0x7ffff7b87406 <interpret_loop+470>    not    %eax                    │ ~insn
│   0x7ffff7b87408 <interpret_loop+472>    test   $0x3,%al                │ (~insn & 3) > 0
│   0x7ffff7b8740a <interpret_loop+474>    jne    0x7ffff7b87360          │ -> decode compressed instruction
// decoding: decode fence.i uncompressed instruction
│   0x7ffff7b87410 <interpret_loop+480>    mov    %ebx,%eax               │
│   0x7ffff7b87412 <interpret_loop+482>    and    $0x707f,%eax            │
│   0x7ffff7b87417 <interpret_loop+487>    cmp    $0x3023,%eax            │
│   0x7ffff7b8741c <interpret_loop+492>    je     0x7ffff7b891e8          │
│   0x7ffff7b87422 <interpret_loop+498>    ja     0x7ffff7b87590          │
│   0x7ffff7b87428 <interpret_loop+504>    cmp    $0x101b,%eax            │
│   0x7ffff7b8742d <interpret_loop+509>    je     0x7ffff7b899f4          │
│   0x7ffff7b87433 <interpret_loop+515>    ja     0x7ffff7b87838          │
│   0x7ffff7b87439 <interpret_loop+521>    cmp    $0x3b,%eax              │
│   0x7ffff7b8743c <interpret_loop+524>    je     0x7ffff7b892b9          │
│   0x7ffff7b87442 <interpret_loop+530>    ja     0x7ffff7b87bc0          │
│   0x7ffff7b87bc0 <interpret_loop+2448>   cmp    $0x1003,%eax            │
│   0x7ffff7b87bc5 <interpret_loop+2453>   je     0x7ffff7b89030          │
│   0x7ffff7b87bcb <interpret_loop+2459>   ja     0x7ffff7b88680          │
│   0x7ffff7b88680 <interpret_loop+5200>   cmp    $0x1013,%eax            │
│   0x7ffff7b88685 <interpret_loop+5205>   je     0x7ffff7b897b0          │
│   0x7ffff7b8868b <interpret_loop+5211>   cmp    $0x1017,%eax            │
│   0x7ffff7b88690 <interpret_loop+5216>   je     0x7ffff7b89317          │
│   0x7ffff7b88696 <interpret_loop+5222>   cmp    $0x100f,%eax            │
│   0x7ffff7b8869b <interpret_loop+5227>   jne    0x7ffff7b89b7a          │
// execute
│   0x7ffff7b886a1 <interpret_loop+5233>   add    $0x4,%rbp               | pc += 4
│   0x7ffff7b886a5 <interpret_loop+5237>   jmp    0x7ffff7b873e0          │ -> jump to begin
```

We can see that in optimal conditions it takes exactly 34 x86_64 instructions to execute one `FENCE.I` in this trace, where:
* mcycle check: 3 instructions (same as before)
* fetch: 5 instructions (-6 instructions)
* decoding: 24 instructions (same as before)
* execution: 2 instructions (same as before)

So in summary 6 instructions were optimized out from the very hot path. Tese are the new numbers for benchmarks:

```
$ lua bench-insns.lua
RISC-V Privileged Memory-management             4.227 MIPS   3244.0 ucycles
RISC-V Privileged Interrupt-management        509.389 MIPS    257.0 ucycles
RV64I - Base integer instruction set          611.176 MIPS    264.3 ucycles
RV64M - Integer multiplication and division   606.949 MIPS    312.9 ucycles
RV64A - Atomic instructions                   449.197 MIPS    304.6 ucycles
RV64F - Single-precision floating-point       227.302 MIPS    489.7 ucycles
RV64D - Double-precision floating-point       225.613 MIPS   1774.1 ucycles
RV64Zicsr - Control and status registers      300.756 MIPS    329.0 ucycles
RV64Zicntr - Base counters and timers         354.946 MIPS    283.3 ucycles
RV64Zifence - Instruction fetch fence         847.884 MIPS    246.0 ucycles

$ hyperfine -w 10 -m 32 "cartesi-machine dhrystone 2000000"
Benchmark 1: cartesi-machine dhrystone 2000000
  Time (mean ± σ):      1.269 s ±  0.011 s    [User: 1.261 s, System: 0.007 s]
  Range (min … max):    1.257 s …  1.317 s    32 runs
```

We can see improvements in all benchmarks, where:
- Dhrystone took about `1.269 /  1.333 = 95%` time to execute, minus 5%.
- `611.176 / 574.50 = ~ 6%` improvement in instruction execution speed for RV64I instruction set

## Round 2 - Optimize decoding for uncompressed instruction
The decoding is using 24 instructions of the 34 instructions in the trace, this is about 70%! It's dominating the hot loop trace, imagine if we cut it in a half, maybe we could with jump tables.

*EDIT:* I decided to give a try to optimize the decoding code in a way so the GCC compiler can optimize it to jump tables. After some thinking and research I added a new commit to this PR, and this is the new trace:

```asm
// mcycle check
|   0x7ffff7b82990 <interpret_loop+560>  add    $0x1,%r14          │ ++mcycle
│   0x7ffff7b82994 <interpret_loop+564>  cmp    %r11,%r14          │ mcycle < mcycle_tick_end
│   0x7ffff7b82997 <interpret_loop+567>  jb     0x7ffff7b828a0     │ -> break interpret hot loop
// fetch
|   0x7ffff7b828a0 <interpret_loop+320>  mov    %r15,%rax          │ pc
│   0x7ffff7b828a3 <interpret_loop+323>  xor    %r13,%rax          │ pc ^ fetch_vaddr_page
│   0x7ffff7b828a6 <interpret_loop+326>  cmp    $0xffd,%rax        │ (pc ^ fetch_vaddr_page) < (PMA_PAGE_SIZE - 2)
│   0x7ffff7b828ac <interpret_loop+332>  ja     0x7ffff7b844f0     │ -> miss fetch cache
│   0x7ffff7b828b2 <interpret_loop+338>  mov    (%r15,%r12,1),%ebx │ insn = *(uint32_t*)(pc + fetch_vh_offset)
// decoding: check if is a compressed instruction
│   0x7ffff7b828b6 <interpret_loop+342>  mov    %ebx,%ecx          │ insn
│   0x7ffff7b828b8 <interpret_loop+344>  and    $0x3,%ecx          │ ~insn
│   0x7ffff7b828bb <interpret_loop+347>  cmp    $0x3,%ecx          │ (~insn & 3) > 0
│   0x7ffff7b828be <interpret_loop+350>  je     0x7ffff7b83100     │ -> decode compressed instruction
// decoding: decode fence.i uncompressed instruction
│   0x7ffff7b83100 <interpret_loop+2464> mov    %ebx,%eax          │ insn
│   0x7ffff7b83102 <interpret_loop+2466> mov    %ebx,%edx          │ insn
│   0x7ffff7b83104 <interpret_loop+2468> shr    $0x5,%eax          │ insn >> 5
│   0x7ffff7b83107 <interpret_loop+2471> and    $0x7f,%edx         │ insn & 0b1111111
│   0x7ffff7b8310a <interpret_loop+2474> and    $0x380,%eax        │ (insn >> 5) & 0b1110000000
│   0x7ffff7b8310f <interpret_loop+2479> or     %edx,%eax          │ ((insn >> 5) & 0b1110000000) | (insn & 0b1111111)
│   0x7ffff7b83111 <interpret_loop+2481> lea    -0x3(%rax),%edx    │ load index in jump table
│   0x7ffff7b83114 <interpret_loop+2484> cmp    $0x3f0,%edx        │ check if index is valid
│   0x7ffff7b8311a <interpret_loop+2490> ja     0x7ffff7b83130     │ -> illegal instruction
│   0x7ffff7b8311c <interpret_loop+2492> lea    0x3b711(%rip),%rdi │ load jump base offset
│   0x7ffff7b83123 <interpret_loop+2499> movslq (%rdi,%rdx,4),%rdx │ load jump offset for given index
│   0x7ffff7b83127 <interpret_loop+2503> add    %rdi,%rdx          │ compute instruction jump address
│   0x7ffff7b8312a <interpret_loop+2506> jmp    *%rdx              │ -> jump to instruction
// execute
│   0x7ffff7b83590 <interpret_loop+3632> add    $0x4,%r15          │ pc += 4
│  >0x7ffff7b83594 <interpret_loop+3636> jmp    0x7ffff7b82990     │ -> jump to begin
```

…