@edubart 2024-12-17
I rebased my optimizations PR on top of Perna's PR, tests passed, but to my surprise the `make test-uarch-compare` test went from 12min to 1hour in CI. I went digging into why, and the culprit was that uarch pristine RAM became way larger with my big jump table, making `reset_uarch` operation heavier. So I had to optimize uarch reset operation to avoid touching unnecessary pages, and now this test is taking 5mins.